#1244 add test before remove log file

+++ scripts/zope_repository_db2.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
-rm zr_db.log
+if [ -e zr_db.log ]; then
+   rm zr_db.log
+fi
 echo "## " `date +"%Y-%m-%d, %H:%M:%S"` " : Start of zope_repository_db" >>zr_db.log
 for i in `cat INSTANCES.txt`
-do /srv/python244/bin/python zope_repository_db.py -i $i >>zr_db.log 2>&1
+do /srv/python246/bin/python zope_repository_db.py -i $i >>zr_db.log 2>&1
 done
 echo "## " `date +"%Y-%m-%d, %H:%M:%S"` " : End of zope_repository_db" >>zr_db.log
